[Refactoring] 라이엇 api DTO의 api마다 패키지 분리
