+++ src/main/java/aresain/loldatastats/riot/dto/match/MatchDto.java
@@ -1,0 +1,11 @@
+package aresain.loldatastats.riot.dto.match;
+
+import aresain.loldatastats.riot.dto.match.InfoDto;
+import aresain.loldatastats.riot.dto.match.MetadataDto;
+import lombok.Getter;
+
+@Getter
+public class MatchDto {
+	MetadataDto metadata;
+	InfoDto info;
+}
