+++ src/main/java/aresain/loldatastats/riot/dto/match/MetadataDto.java
@@ -1,0 +1,14 @@
+package aresain.loldatastats.riot.dto.match;
+
+import java.util.List;
+
+import lombok.Getter;
+
+@Getter
+public class MetadataDto {
+	private String dataVersion;
+	private String matchId;
+	private List<String> participants;
+}
+
+
